+++ app.mjs
@@ -12,7 +12,7 @@ const io = new Server(server)
 app.use('/css', express.static('./static/css'))
 app.use('/js', express.static('./static/js'))
 
-// /opentalk 경로로 요청이 들어왔을 때는 router 미들웨어를 사용합니다.
+// /opentalk 경로로 요청이 들어왔을 때는 router 미들웨어 사용
 app.use('/opentalk', router)
 
 router.get('/', (request, response) => {
