+++ src/components/Footer.vue
@@ -21,25 +21,4 @@ const toggleLocales = () => {
       CopyRight 2023. Reeltalks team. All right reserved.
     </p>
   </div>
-  <nav text-xl mt-6>
-    <RouterLink class="icon-btn mx-2" to="/" :title="t('button.home')">
-      <div i-carbon-campsite />
-    </RouterLink>
-
-    <button class="icon-btn mx-2 !outline-none" :title="t('button.toggle_dark')" @click="toggleDark()">
-      <div i="carbon-sun dark:carbon-moon" />
-    </button>
-
-    <a class="icon-btn mx-2" :title="t('button.toggle_langs')" @click="toggleLocales()">
-      <div i-carbon-language />
-    </a>
-
-    <RouterLink class="icon-btn mx-2" to="/guide" :title="t('button.guide')">
-      <div i-carbon-dicom-overlay />
-    </RouterLink>
-
-    <a class="icon-btn mx-2" rel="noreferrer" href="https://github.com/socketbear/vue-dev-guide" target="_blank" title="GitHub">
-      <div i-carbon-logo-github />
-    </a>
-  </nav>
 </template>
