+++ src/components/chat/TalkWindow.vue
@@ -1,0 +1,11 @@
+<script setup>
+// TT
+</script>
+
+<template>
+  <div>
+    dfdf
+  </div>
+</template>
+
+<style lang="scss" scoped></style>
