+++ src/layouts/default.vue
@@ -1,9 +1,10 @@
 <template>
-  <main class="px-4 py-10 text-center text-gray-700 dark:text-gray-200">
-    <RouterView />
+  <div>
+    <Header />
+    <main class="px-4 py-10 text-center text-gray-700 dark:text-gray-200">
+      <RouterView />
+      <div class="mt-5 mx-auto text-center opacity-75 dark:opacity-50 text-sm" />
+    </main>
     <Footer />
-    <div class="mt-5 mx-auto text-center opacity-75 dark:opacity-50 text-sm">
-      [Default Layout]
-    </div>
-  </main>
+  </div>
 </template>
