+++ src/main.ts
@@ -1,5 +1,6 @@
 import { ViteSSG } from 'vite-ssg'
 import { setupLayouts } from 'virtual:generated-layouts'
+
 import App from './App.vue'
 import type { UserModule } from './types'
 import generatedRoutes from '~pages'
