+++ src/pages/sandbox.vue
@@ -1,0 +1,3 @@
+<template>
+  <chat-talk-window />
+</template>
